Reorganize Graph/Info tabs
