+++ components/AboutPane/ColorizeInput.scss
