+++ components/AboutPane/GithubSponsorButton.scss
@@ -1,7 +1,7 @@
 .pane a.github-sponsor {
   display: inline-block;
 
-  margin-left: 1rem;
+  margin-block: 1rem;
 
   border: solid 1px var(--bg1);
   border-radius: 0.375rem;
