+++ components/GraphPane/GraphPane.scss
@@ -6,8 +6,6 @@
 
 #colorize-ui {
   margin-top: 1em;
-  border-top: solid 1px var(--bg1);
-  padding-top: 1em;
 }
 
 #colorize-ui select {
