+++ components/GraphPane/GraphPane.tsx
@@ -1,15 +1,7 @@
 import React from 'react';
-import simplur from 'simplur';
-import useHashParam from '../../lib/useHashParam.js';
-
-import { PARAM_DEPENDENCIES, PARAM_SIZING } from '../../lib/constants.js';
-import { isDefined } from '../../lib/guards.js';
-import useCollapse from '../../lib/useCollapse.js';
 import { ExternalLink } from '../ExternalLink.js';
-import { DependencyKey, GraphState } from '../GraphDiagram/graph_util.js';
+import { GraphState } from '../GraphDiagram/graph_util.js';
 import { Pane } from '../Pane.js';
-import { Toggle } from '../Toggle.js';
-import ColorizeInput from './ColorizeInput.js';
 import './GraphPane.scss';
 import { ReportItem } from './reports/ReportItem.js';
 import { analyzeLicenses } from './reports/analyzeLicenses.js';
@@ -33,15 +25,6 @@ export default function GraphPane({
   graph,
   ...props
 }: { graph: GraphState | null } & React.HTMLAttributes<HTMLDivElement>) {
-  const [collapse, setCollapse] = useCollapse();
-  const [depTypes, setDepTypes] = useHashParam(PARAM_DEPENDENCIES);
-  const [sizing, setSizing] = useHashParam(PARAM_SIZING);
-
-  const dependencyTypes = (
-    (depTypes ?? '').split(/\s*,\s*/) as DependencyKey[]
-  ).filter(isDefined);
-
-  const includeDev = dependencyTypes.includes('devDependencies');
   if (!graph?.moduleInfos) return <div>Loading</div>;
 
   const moduleAnalysis = analyzeModules(graph);
@@ -51,41 +34,6 @@ export default function GraphPane({
 
   return (
     <Pane {...props}>
-      <Toggle
-        checked={includeDev}
-        style={{ marginTop: '1rem' }}
-        onChange={() => setDepTypes(includeDev ? '' : 'devDependencies')}
-      >
-        Include devDependencies
-      </Toggle>
-
-      <Toggle
-        checked={sizing === ''}
-        style={{ marginTop: '1rem' }}
-        onChange={() => setSizing(sizing === null)}
-      >
-        Size modules by unpacked size
-      </Toggle>
-
-      <ColorizeInput />
-
-      <div
-        style={{
-          fontSize: '90%',
-          color: 'var(--text-dim)',
-          marginTop: '1em',
-        }}
-      >
-        {collapse.length ? (
-          <span>
-            {simplur`${collapse.length} module[|s] collapsed `}
-            <button onClick={() => setCollapse([])}>Expand All</button>
-          </span>
-        ) : (
-          <span>(Shift-click modules in graph to expand/collapse)</span>
-        )}
-      </div>
-
       <h3>Modules</h3>
 
       <ReportItem data={moduleAnalysis} reporter={modulesAll} />
