+++ components/Inspector.scss
@@ -87,4 +87,7 @@ label {
   a {
     color: var(--highlight);
   }
+  hr {
+    margin-inline: -1em;
+  }
 }
